Make tests concurrency-safe
The tests now all operate on a temporary directory, potentially
containing the contents of the "testdata" directory. The contents is
not copied as-is, but destination filenames are derived via
percent-decoding from the source filenames.

This allows the tests to safely run concurrently, so the flags
enforcing single-threaded execution have been removed from the CI
config.

Extending the API slightly, this commit also adds a `Maildir::path`
method, which came handy when refactoring the tests.

Fixes #6.