store_new and store_cur_with_flags now return the id of the inserted mail
Previously there was no trivial way how one could get the id of a freshly
inserted email. This commit modifies the return type of Maildir::store to
Result<String, MaildirError>, and returns the id on success.

The smoke test for store_new was modified to verify that the id is actually the
correct one.