fix TOCTOU when racing with filenames, improve cleanup
Avoid using path.exists() in a loop as it is prone to TOCTOU
when racing with other actors.

Instead, use the create_new option when opening the file for
read so that the kernel can handle and advise of races.

When looping, rather than bumping the timestamp, fetch the
now-current timestamp and instead maintain and increment a
process-local counter and include that in the candidate id.

If the final rename fails, remove the temporary file that
we created.